Remove any license restrictions.
Every file in this repository is now dedicated to the public domain, under the Unlicense "license".

I dedicate any and all copyright interest in this software to the
public domain. I make this dedication for the benefit of the public at
large and to the detriment of my heirs and successors. I intend this
dedication to be an overt act of relinquishment in perpetuity of all
present and future rights to this software under copyright law.

See also: https://unlicense.org/